+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This project uses [Semantic Versioning 2.0.0](http://semver.org/), the format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+### Deprecated
+
+- Deprecated `ParentId` in `ZoneRecord`. The value is always null, and the property will be removed in the next major version. (#263)
+
 ## 2.0.0 - 2026-05-05
 
 ### Changed